// wardn/frontend/lib/api/generated/model/invitationPreview.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.0.1
 */
import type { InvitationPreviewAuthMode } from './invitationPreviewAuthMode';
import type { InvitationPreviewScopeType } from './invitationPreviewScopeType';

export interface InvitationPreview {
  authMode: InvitationPreviewAuthMode;
  currentUserEmail?: string | null;
  email: string;
  expiresAt: string;
  oidcProviderName: string;
  organizationId: string;
  organizationName: string;
  role: string;
  scopeType: InvitationPreviewScopeType;
  workspaceId: string | null;
  workspaceName: string | null;
}